Move cmd handlng to separate file

+++ plug/src/util.rs
@@ -4,8 +4,8 @@ pub trait Narrow<T> {
     fn narrow( self, above: T, below: T) -> T;
 }
 
-impl Narrow<f32> for f32 {
-    fn narrow( self, above: f32, below: f32) -> f32 {
+impl Narrow<f64> for f64 {
+    fn narrow( self, above: f64, below: f64) -> f64 {
         match self {
             a if a < above => above,
             a if a > below => below,
